requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
python-multipart
aiofiles
aiohttp
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
